# crAPIEncryption/api-specs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OWASP crAPI API",
    "version": "1.0"
  },
...
  "paths": {
...
    "/community/api/v2/coupon/new-coupon": {
      "post": {
        "operationId": "OP_COMMUNITY_COUPON_ADD_COUPON",
        "summary": "Add a new coupon to the shop database",
        "tags": [
          "Community / Coupon"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_MECHANIC",
          "ROLE_ADMIN"
        ],
        "requestBody": {
          "description": "Coupon",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Coupon"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Coupon Added in database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/community/api/v2/coupon/validate-coupon": {
      "post": {
        "operationId": "OP_COMMUNITY_COUPON_VALIDATE_COUPON",
        "summary": "Validate the provided discount coupon code",
        "tags": [
          "Community / Coupon"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_USER",
          "ROLE_MECHANIC",
          "ROLE_ADMIN"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "coupon_code"
                ],
                "properties": {
                  "coupon_code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "CreatedAt",
                    "amount",
                    "coupon_code"
                  ],
                  "properties": {
                    "CreatedAt": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "string"
                    },
                    "coupon_code": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "OP_COMMUNITY_COUPON_GET_VALIDATE_COUPON",
        "summary": "Validate the provided discount coupon code",
        "tags": [
          "Community / Coupon"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_USER",
          "ROLE_MECHANIC",
          "ROLE_ADMIN"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "coupon_code",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "TRAC075"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "CreatedAt",
                    "amount",
                    "coupon_code"
                  ],
                  "properties": {
                    "CreatedAt": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "string"
                    },
                    "coupon_code": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": []
    },
    "/community/api/v1/coupon/validate-coupon": {
      "get": {
        "operationId": "OP_COMMUNITY_COUPON_GET_V1_VALIDATE_COUPON",
        "summary": "Validate the provided discount coupon code",
        "tags": [
          "Community / Coupon"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_USER",
          "ROLE_MECHANIC",
          "ROLE_ADMIN"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "coupon_code",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "TRAC075"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "CreatedAt",
                    "amount",
                    "coupon_code"
                  ],
                  "properties": {
                    "CreatedAt": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "string"
                    },
                    "coupon_code": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": []
    },
...
  },
  "components": {
    "schemas": {
...
      "Coupon": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string"
          },
          "coupon_code": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "amount",
          "coupon_code"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}